requirements.txt
PyQt6>=6.4.0
requests>=2.28.0
google-generativeai>=0.3.0
pynput>=1.7.0
mss>=7.0.0
Pillow>=9.0.0
pyperclip>=1.8.0
python-dotenv>=0.20.0
